+++ examples/utils/io.py
@@ -4,7 +4,7 @@ from typing import Optional
 from pacha.data_engine.data_engine import SqlHooks
 from pacha.query_planner.query_planner import QueryPlannerHooks, QueryPlan, DataContext
 from pacha.data_engine import SqlOutput
-from pacha.query_planner.python_executor import PythonExecutorHooks
+from pacha.data_engine.python_executor import PythonExecutorHooks
 
 ENDC = '\033[0m'
